refactor(bi): move source files into nested python package
This should make the files easier to debug, and allow mypy to properly lint the project.
