+++ data/books/cs6e/ch07/fig0738/lang/mnemonics.py
